Create AppFormSearch.vue
